refactor(speech): address review findings on the state_machine split
- Qualify the three [run_loop] intra-doc links in types.rs as
  [super::run_loop] — they resolved in the old flat file but broke
  when run_loop moved to the sibling entry.rs (cargo doc emitted 3
  new unresolved-link warnings; now 0).
- Move handle_llm_err into inner.rs beside its only call sites,
  ending the parent-holds-child's-helper inversion; mod.rs keeps only
  FALLBACK_LINE (which mocks.rs asserts against via super::), with a
  comment stating that constraint.
- Drop the #[cfg(test)]-gated observer-type re-import from mod.rs;
  mocks.rs now imports ExitReason/LoopObserver/TurnCompletePayload/
  VoiceState directly from voice_loop::observer, so imports live in
  the file that uses them.
- Update the relocated NOTE comment in mod.rs ("This file owns the
  state machine…" was false post-split) to describe the submodule
  layout.
- Correct the CLAUDE.md layout note (mocks.rs never reached
  handle_llm_err via super::, only FALLBACK_LINE) and refresh
  NEXT_SESSION.md to match.

Verification: cargo test -p primer-speech --features voice-loop
(101 + 1 + 10, identical baseline); cargo test --workspace (51 ok /
0 failed); clippy -D warnings clean on default and
voice-loop,silero,whisper,piper arms; fmt clean; cargo doc shows the
3 run_loop link warnings gone.
