+++ NEXT_SESSION.md
@@ -9,10 +9,11 @@ This session ran the standard sweep FIRST and the inline-test detector **fired f
 
 - **PR #317 (branch `refactor/extract-comprehension-llm-tests`), one commit `ebcccf4`** — `refactor(comprehension): extract llm tests to sibling file`. Moves the inline `#[cfg(test)] mod tests` out of `primer-comprehension/src/llm.rs` into a sibling `llm/tests.rs` (non-root module → one dir deeper, mkdir), following the `macos26/analyzer` precedent (#295): 717 → 317 (production) + 399 (tests). Verbatim relocation plus the rustfmt re-wraps triggered by the one-level dedent; the test module stays a child of `llm`, so `use super::*` still reaches the private `EXAMPLE_OUTPUT` with no visibility edits. **20 lib tests identical pre/post**; clippy `-D warnings` + fmt clean; pub-surface diff empty; no cargo features on the crate. **CI: all 11 checks green.**
 - **PR #318 (branch `refactor/split-voice-loop-state-machine`), one commit `901b1a7`** — `refactor(speech): split voice_loop state_machine.rs into responsibility submodules`. The 786-line `state_machine.rs` became a directory-style module (the dir already existed for `mocks.rs`). All files < 500 lines:
-  - `state_machine/mod.rs` (74) — module doc, flat `pub use` façade (9-symbol external surface byte-identical: `DrainHook, LoopBackends, LoopConfig, LoopHandle, Responder, VAD_EVENT_CHANNEL_CAPACITY, VoiceLoopError, run_loop, run_loop_borrowed`), and the private `FALLBACK_LINE` + `handle_llm_err`. These stay in `mod.rs` because `mocks.rs` (the test module, a child of this module) reaches `FALLBACK_LINE` via `super::`; a `#[cfg(test)]`-gated `use super::observer::{ExitReason, TurnCompletePayload, VoiceState};` keeps mocks' `super::X` paths resolving — **zero churn in the 1380-line mocks.rs**.
-  - `state_machine/types.rs` (189) — `LoopConfig`, `LoopBackends` (+ impl), `Responder`, `LoopHandle`, `DrainHook`, `VoiceLoopError`, `VAD_EVENT_CHANNEL_CAPACITY` (verbatim).
+  - `state_machine/mod.rs` — module doc, flat `pub use` façade (9-symbol external surface byte-identical: `DrainHook, LoopBackends, LoopConfig, LoopHandle, Responder, VAD_EVENT_CHANNEL_CAPACITY, VoiceLoopError, run_loop, run_loop_borrowed`), and the private `FALLBACK_LINE` — kept there because `mocks.rs` (the test module, a child of this module) asserts against it via `super::FALLBACK_LINE`.
+  - `state_machine/types.rs` (189) — `LoopConfig`, `LoopBackends` (+ impl), `Responder`, `LoopHandle`, `DrainHook`, `VoiceLoopError`, `VAD_EVENT_CHANNEL_CAPACITY` (verbatim; the `[run_loop]` intra-doc links qualified to `[super::run_loop]` post-review so `cargo doc` stays warning-free).
   - `state_machine/entry.rs` (90) — `run_loop` + `run_loop_borrowed` (verbatim).
-  - `state_machine/inner.rs` (470) — `run_loop_inner`, now `pub(super)`. Only non-verbatim change: imports — observer/quit_detect/tts_markdown via `crate::voice_loop::…` paths (the old flat file used parent-relative `super::…`, which would have become `super::super::…`).
+  - `state_machine/inner.rs` — `run_loop_inner`, now `pub(super)`, plus the private `handle_llm_err` helper (moved next to its only call sites in the post-review pass). Imports — observer/quit_detect/tts_markdown via `crate::voice_loop::…` paths (the old flat file used parent-relative `super::…`, which would have become `super::super::…`).
+  - `state_machine/mocks.rs` — one-line import churn from the review pass: the observer types (`ExitReason`/`LoopObserver`/`TurnCompletePayload`/`VoiceState`) are now imported directly from `crate::voice_loop::observer` instead of resolving through a `#[cfg(test)]`-gated re-import in `mod.rs` (imports live in the file that uses them).
   - CLAUDE.md got a module-layout note appended to the `primer-speech` bullet (mirrors the #304/#305 precedent). README/ROADMAP untouched (internal refactors; verified no drift — neither file mentions `state_machine`/`run_loop`).
 
 ### Verification
